+++ notebooks/disable_ops.py
@@ -15,10 +15,9 @@
 # %%
 import torch
 
-from aioway.modes import mode_off
-
 # %%
-from aioway.tracking import PrintTorchDisp, PrintTorchFunc
+from aioway.ctx import PrintTorchDisp, PrintTorchFunc
+from aioway.modes import mode_off
 
 # %%
 a = torch.randn(3, 4)
