+++ notebooks/fate.py
@@ -16,9 +16,9 @@
 import torch
 
 from aioway._torch import *
+from aioway.ctx import *
 from aioway.fate import *
 from aioway.modes import *
-from aioway.tracking import *
 
 # %%
 dispatch_print = PrintTorchDisp()
