+++ notebooks/print_torch_ops.py
@@ -17,7 +17,7 @@ import torch
 
 # %%
 from aioway._torch import torch_fake_mode
-from aioway.tracking import PrintTorchDisp, PrintTorchFunc
+from aioway.ctx import PrintTorchDisp, PrintTorchFunc
 
 # %%
 a = torch.randn(3, 4)
