+++ tests/conftest.py
@@ -10,7 +10,7 @@ import torch
 from numpy import random as npr
 from rich import traceback
 
-from aioway.tracking import fake_fn, track_fn
+from aioway.ctx import fake_fn, track_fn
 
 from .fake import batch_sizes, cpu_and_maybe_cuda
