+++ tests/modes/test_history.py
@@ -5,8 +5,8 @@ import torch
 from torch import nn
 
 from aioway._torch import current_fake_mode, torch_fake_mode
+from aioway.ctx import fake_fn, track_fn
 from aioway.modes import NnFwdFn, NnInitFn
-from aioway.tracking import fake_fn, track_fn
 
 
 @pytest.fixture
